refactor(primitives): move compressed array to record-batch and rename
